feat: Add Badge when documents downloaded while offline - EXO-79902 - EXIP-15 This change will display a badge in Documents App when the user had downloaded documents while the site is offline.
