--- documents-webapp/src/main/webapp/vue-app/documents-favorite-drawer-extensions/components/DocumentsFavoriteOfflineButton.vue
@@ -1,58 +1,0 @@
-<!--
-
-  Copyright (C) 2003 - 2025 eXo Platform SAS.
-
-  This program is free software: you can redistribute it and/or modify
-  it under the terms of the GNU Affero General Public License
-  as published by the Free Software Foundation, either version 3 of the License, or
-  (at your option) any later version.
-
-  This program is distributed in the hope that it will be useful,
-  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-  GNU Affero General Public License for more details.
-
-  You should have received a copy of the GNU Affero General Public License
-  along with this program. If not, see <gnu.org/licenses>.
-
--->
-<template>
-  <div
-    v-if="offlineModeEnabled"
-    :title="hasDocmentsAccessedOffline ? $t('OfflineApp.pwa.documents.uploadDocumentsTooltip') : $t('OfflineApp.pwa.documents.accessDocumentsTooltip')">
-    <v-badge
-      :value="hasDocmentsAccessedOffline"
-      color="var(--allPagesBadgePrimaryColor, #d32a2a)"
-      offset-x="12"
-      offset-y="12"
-      overlap
-      flat
-      dot>
-      <v-btn
-        id="offlineDocumentsButton"
-        class="ms-4"
-        icon
-        @click="$root.$emit('open-document-offline-files')">
-        <v-icon size="20">fa-power-off</v-icon>
-      </v-btn>
-    </v-badge>
-    <documents-offline-drawer />
-  </div>
-</template>
-<script>
-export default {
-  data: () => ({
-    offlineModeEnabled: false,
-    hasDocmentsAccessedOffline: false,
-  }),
-  created() {
-    this.init();
-  },
-  methods: {
-    async init() {
-      this.offlineModeEnabled = await this.$documentOfflineService.isDatabaseExists();
-      this.hasDocmentsAccessedOffline = await this.$documentOfflineService.hasDocumentsAccessedOffline();
-    },
-  }
-};
-</script>
